fix #3 and add retry button

+++ src/app/shared/service/udemy.service.ts
@@ -9,7 +9,7 @@ import {Lecture} from '../model/lecture.model';
 import {ElectronService} from 'ngx-electron';
 import * as sanitize from 'sanitize-filename';
 import {AssetType} from '../model/asset-type.model';
-import {concatMap, filter, flatMap, map, retryWhen} from 'rxjs/operators';
+import {concatMap, filter, flatMap, map, retryWhen, tap} from 'rxjs/operators';
 import {MatSnackBar} from '@angular/material';
 import {environment} from '../../../environments/environment';
 import {WriteStream} from 'fs';
@@ -17,6 +17,7 @@ import {SettingsService} from './settings.service';
 import {DownloadProgress} from '../model/download-progress.model';
 import {FileMetadata} from '../model/file-metadata.model';
 import {VideoQuality} from '../model/video-quality.model';
+import {CourseMetadata} from '../model/course-metadata.model';
 
 class AssetDownloadable {
   constructor(
@@ -74,7 +75,7 @@ export class UdemyService {
   getSubscribedCourses(): Observable<Result<Course>> {
     return this.http.get<Result<Course>>(
         // tslint:disable-next-line
-        `https://${this.auth.getSubDomain()}.udemy.com/api-2.0/users/me/subscribed-courses?page_size=50&fields[course]=@default,completion_ratio,visible_instructors`,
+        `https://${this.auth.getSubDomain()}.udemy.com/api-2.0/users/me/subscribed-courses?page_size=100&fields[course]=@default,completion_ratio,visible_instructors`,
         {
           headers: this.authHeaders
         }
@@ -134,41 +135,48 @@ export class UdemyService {
   }
 
   downloadAsset(url: string, path: string): Observable<AssetMetadata<ArrayBuffer>> {
-    return this.http.get(url,
-        {
-          headers: this.downloadHeaders,
-          reportProgress: true,
-          responseType: 'arraybuffer',
-          // observe: 'events'
-          observe: 'response'
-        })
+    return this.http.head(url)
     .pipe(
-        retryWhen(() => {
-          return interval(environment.download.interval).pipe(
-              flatMap(count => count === environment.download.retry ? throwError(`Error during download of ${url} for ${path}`) : of(count))
+        flatMap(() => {
+          return this.http.get(url,
+              {
+                headers: this.downloadHeaders,
+                reportProgress: true,
+                responseType: 'arraybuffer',
+                // observe: 'events'
+                observe: 'response'
+              })
+          .pipe(
+              retryWhen(() => {
+                return interval(environment.download.interval).pipe(
+                    flatMap(count =>
+                        count === environment.download.retry ? throwError(`Error during download of ${url} for ${path}`) : of(count))
+                );
+              }),
+              map(value => new AssetMetadata<ArrayBuffer>(path, value.body))
           );
-        }),
-        map(value => new AssetMetadata<ArrayBuffer>(path, value.body))
+        })
     );
+
   }
 
-  downloadArticle(courseId: number, lectureId: number, dir: string, lectureIdx: number): Observable<AssetMetadata<String>> {
+  downloadArticle(courseId: number, lectureId: number, path: string, lectureIdx: number): Observable<AssetMetadata<String>> {
     return this.getLecture(courseId, lectureId)
     .pipe(
         map((lecture: Lecture) => {
           return <AssetMetadata<String>>{
-            path: `${dir}/${this.numberOptimization(lectureIdx)} - ${sanitize(lecture.title)}.html`,
+            path: `${path}/${this.numberOptimization(lectureIdx)} - ${sanitize(lecture.title)}.html`,
             data: lecture.asset.body
           };
         }),
     );
   }
 
-  getVideoDownloadable(courseId: number, lectureId: number, dir: string, lectureIdx: number): Observable<AssetDownloadable> {
+  getVideoDownloadable(courseId: number, lectureId: number, path: string, lectureIdx: number): Observable<AssetDownloadable> {
     return this.getLecture(courseId, lectureId)
     .pipe(
         map((lecture: Lecture) => {
-          let filePath: string = `${dir}/${this.numberOptimization(lectureIdx)} - ${sanitize(lecture.title)}`;
+          let filePath: string = `${path}/${this.numberOptimization(lectureIdx)} - ${sanitize(lecture.title)}`;
           const extIdx: number = lecture.asset.filename.lastIndexOf('.');
           const ext: string = lecture.asset.filename.slice(extIdx, lecture.asset.filename.length);
           const fileUrl: string = this.selectVideo(lecture.asset.download_urls.Video).file;
@@ -177,8 +185,10 @@ export class UdemyService {
           if (lecture.supplementary_assets && lecture.supplementary_assets.length > 0) {
             lecture.supplementary_assets.forEach(asset => {
               if (asset.asset_type === AssetType.FILE) {
-                const path: string = `${dir}/${this.numberOptimization(lectureIdx)} - ${sanitize(asset.title)}`;
-                downloadable.assets.push(new DownloadableAssetMetadata(asset.download_urls.File[0].file, path));
+                downloadable.assets.push(new DownloadableAssetMetadata(
+                    asset.download_urls.File[0].file,
+                    `${path}/${this.numberOptimization(lectureIdx)} - ${sanitize(asset.title)}`
+                ));
               } else {
                 console.log('Unknown supplementary_assets');
                 console.log(asset);
@@ -190,11 +200,11 @@ export class UdemyService {
     );
   }
 
-  getEBookDownloadable(courseId: number, lectureId: number, dir: string, lectureIdx: number): Observable<AssetDownloadable> {
+  getEBookDownloadable(courseId: number, lectureId: number, path: string, lectureIdx: number): Observable<AssetDownloadable> {
     return this.getLecture(courseId, lectureId)
     .pipe(
         map((lecture: Lecture) => {
-          const filePath: string = `${dir}/${this.numberOptimization(lectureIdx)} - ${sanitize(lecture.asset.title)}`;
+          const filePath: string = `${path}/${this.numberOptimization(lectureIdx)} - ${sanitize(lecture.asset.title)}`;
           const fileUrl: string = lecture.asset.download_urls['E-Book'][0].file;
           return new AssetDownloadable(new DownloadableAssetMetadata(fileUrl, filePath));
         }),
@@ -207,36 +217,36 @@ export class UdemyService {
     const downloadProgress: BehaviorSubject<DownloadProgress> = new BehaviorSubject<DownloadProgress>(null);
     const downloadSaveAssets: BehaviorSubject<DownloadableAssetMetadata> = new BehaviorSubject<DownloadableAssetMetadata>(null);
     const saveArticles: BehaviorSubject<AssetMetadata<String>> = new BehaviorSubject<AssetMetadata<String>>(null);
-    const courseDir: string = `${this.settingsService.getDownloadPath()}/${sanitize(title)}`;
-    if (this.fs.existsSync(courseDir)) {
-      console.log(`Directory already exist ${courseDir}`);
+    const coursePath: string = this.getCoursePath(title);
+    if (this.fs.existsSync(coursePath)) {
+      console.log(`Directory already exist ${coursePath}`);
     } else {
-      this.fs.mkdirSync(courseDir);
-      console.log(`Create directory ${courseDir}`);
+      this.fs.mkdirSync(coursePath);
+      console.log(`Create directory ${coursePath}`);
     }
     if (imageUrl !== null) {
       totalFiles++;
       const extIdx: number = imageUrl.lastIndexOf('.');
       const ext: string = imageUrl.slice(extIdx, imageUrl.length);
-      downloadSaveAssets.next(new DownloadableAssetMetadata(imageUrl, `${courseDir}/logo${ext}`));
+      downloadSaveAssets.next(new DownloadableAssetMetadata(imageUrl, `${coursePath}/logo${ext}`));
     }
     this.getCourse(id)
     .subscribe(course => {
-      let chapterDir: string;
+      let chapterPath: string;
       let chapterIdx: number = 1;
       let lectureIdx: number = 1;
       if (course.results[0]._class === 'lecture') {
-        chapterDir = `${courseDir}/${this.numberOptimization(chapterIdx++)} - ${sanitize('Course Introduction')}`;
-        if (!this.fs.existsSync(chapterDir)) {
-          this.fs.mkdirSync(chapterDir);
+        chapterPath = `${coursePath}/${this.numberOptimization(chapterIdx++)} - ${sanitize('Course Introduction')}`;
+        if (!this.fs.existsSync(chapterPath)) {
+          this.fs.mkdirSync(chapterPath);
         }
       }
-      course.results.forEach(block => {
+      course.results.forEach((block) => {
             switch (block._class) {
               case 'chapter' :
-                chapterDir = `${courseDir}/${this.numberOptimization(chapterIdx++)} - ${sanitize(block.title)}`;
-                if (!this.fs.existsSync(chapterDir)) {
-                  this.fs.mkdirSync(chapterDir);
+                chapterPath = `${coursePath}/${this.numberOptimization(chapterIdx++)} - ${sanitize(block.title)}`;
+                if (!this.fs.existsSync(chapterPath)) {
+                  this.fs.mkdirSync(chapterPath);
                 }
                 lectureIdx = 1;
                 break;
@@ -249,7 +259,7 @@ export class UdemyService {
                     .filter(value => value.asset_type === AssetType.FILE)
                     .forEach(() => totalFiles++);
                   }
-                  this.getVideoDownloadable(id, lecture.id, chapterDir, lectureIdx++)
+                  this.getVideoDownloadable(id, lecture.id, chapterPath, lectureIdx++)
                   .subscribe(value => {
                     if (value.main) {
                       downloadSaveAssets.next(value.main);
@@ -258,7 +268,7 @@ export class UdemyService {
                   });
                 } else if (lecture.asset.asset_type === AssetType.E_BOOK) {
                   totalFiles++;
-                  this.getEBookDownloadable(id, lecture.id, chapterDir, lectureIdx++)
+                  this.getEBookDownloadable(id, lecture.id, chapterPath, lectureIdx++)
                   .subscribe(value => {
                     if (value.main) {
                       downloadSaveAssets.next(value.main);
@@ -266,7 +276,7 @@ export class UdemyService {
                   });
                 } else if (lecture.asset.asset_type === AssetType.ARTICLE) {
                   totalFiles++;
-                  this.downloadArticle(id, lecture.id, chapterDir, lectureIdx++)
+                  this.downloadArticle(id, lecture.id, chapterPath, lectureIdx++)
                   .subscribe(value => saveArticles.next(value));
                 } else {
                   console.log('unknown lecture');
@@ -288,8 +298,12 @@ export class UdemyService {
         filter(value => value != null)
     )
     .subscribe(value => {
-      this.fs.writeFileSync(value.path, value.data);
-      console.log(`Saved lecture : ${value.path}`);
+      if (this.fs.existsSync(value.path)) {
+        console.log(`File already exist : ${value.path}`);
+      } else {
+        this.fs.writeFileSync(value.path, value.data);
+        console.log(`Saved lecture : ${value.path}`);
+      }
       downloadProgress.next(new DownloadProgress(totalFiles, ++currentFile));
     });
     // Download and Save Assets
@@ -332,26 +346,29 @@ export class UdemyService {
           console.log(error);
           this.snackBar.open(
               `'${title}' Error during downloading.`,
-              '',
-              {
-                duration: environment.message.duration
-              }
+              'Close'
           );
           downloadProgress.next(new DownloadProgress(totalFiles, currentFile, true));
+          this.saveCourseMetadata(title, {complete: false});
         },
         () => {
           console.log(`'${title}' Download completed.`);
           this.snackBar.open(
               `'${title}' Download completed.`,
-              '',
-              {
-                duration: environment.message.duration
-              }
+              'Close'
           );
+          this.saveCourseMetadata(title, {complete: true});
         });
     return downloadProgress.asObservable()
     .pipe(
-        filter(value => value != null)
+        filter(value => value != null),
+        tap(next => {
+          // skip logo file
+          if (next.totalFiles > 1 && next.isDone) {
+            downloadSaveAssets.complete();
+            saveArticles.complete();
+          }
+        })
     );
   }
 
@@ -360,5 +377,22 @@ export class UdemyService {
       return `0${idx}`;
     }
     return idx.toString();
+  }
+
+  private getCoursePath(title: string): string {
+    return `${this.settingsService.getDownloadPath()}/${sanitize(title)}`;
+  }
+
+  getCourseMetadata(title: string): CourseMetadata | null {
+    const metadataPath: string = `${this.getCoursePath(title)}/metadata.json`;
+    if (this.fs.existsSync(metadataPath)) {
+      return JSON.parse(this.fs.readFileSync(metadataPath));
+    }
+    return null;
+  }
+
+  saveCourseMetadata(title: string, metadata: CourseMetadata): void {
+    const metadataPath: string = `${this.getCoursePath(title)}/metadata.json`;
+    this.fs.writeFileSync(metadataPath, JSON.stringify(metadata));
   }
 }
